Move Login.tsx to a view directory
