+++ desktop/packages/mullvad-vpn/src/renderer/components/AppRouter.tsx
@@ -2,8 +2,8 @@ import { useCallback, useRef } from 'react';
 import { Route, Switch } from 'react-router';
 
 import { RoutePath } from '../../shared/routes';
-import LoginPage from '../components/Login';
 import SelectLocation from '../components/select-location/SelectLocationContainer';
+import LoginPage from '../components/views/login/Login';
 import { useViewTransitions } from '../lib/transition-hooks';
 import Account from './Account';
 import ApiAccessMethods from './ApiAccessMethods';
